pyproject.toml
[project]
name = "api-yuri"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "faker>=37.4.0",
    "fastapi>=0.115.14",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.34.3",
]
